fix: bugs found during playtest

+++ Arena/ArenaOnlineGameModes/TeamBattle/TeamBattle.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 using ArenaMode = RainMeadow.ArenaOnlineGameMode;
@@ -345,7 +345,10 @@ namespace RainMeadow.Arena.ArenaOnlineGameModes.TeamBattle
                         }
 
                         if (aIsWinningTeam && bIsWinningTeam)
-                        {
+                        { 
+                            if (A.alive != B.alive) {
+                                return A.alive;
+                            }
                             if (A.score != B.score)
                             {
                                 return A.score > B.score; // If both are on winning team, sort by kill value
@@ -355,10 +358,10 @@ namespace RainMeadow.Arena.ArenaOnlineGameModes.TeamBattle
                             {
                                 return A.allKills.Count > B.allKills.Count;
                             }
-                            else if (A.deaths != B.deaths)
-                            {
-                                return A.deaths < B.deaths;
-                            }
+                            
+                            
+                            return A.deaths < B.deaths;
+                            
                         }
                     }
                 }
